Move NetworkingClientTestInterceptor to own file
