+++ Tests/StytchCoreTests/TOTPTestCase.swift
@@ -48,53 +48,3 @@ final class TOTPTestCase: BaseTestCase {
     }
 }
 
-import Foundation
-
-final class NetworkingClientInterceptor {
-    var requests: [URLRequest] = []
-    var responses: [Result<Data, Swift.Error>] = []
-
-    func reset() {
-        requests = []
-        responses = []
-    }
-
-    func handleRequest(request: URLRequest) async throws -> (Data, HTTPURLResponse) {
-        requests.append(request)
-
-        switch responses.removeFirst() {
-        case let .success(data):
-            return try (
-                data,
-                XCTUnwrap(.init(url: XCTUnwrap(request.url), statusCode: 200, httpVersion: "HTTP/1.1", headerFields: [:]))
-            )
-        case let .failure(error):
-            throw error
-        }
-    }
-
-    func appendSuccess(data: Data...) {
-        responses.append(contentsOf: data.map { .success($0.surroundInDataJSONContainer()) })
-    }
-
-    func appendSuccess<T: Codable>(_ elements: T...) throws {
-        responses.append(
-            contentsOf: try elements.map {
-                .success(try Current.jsonEncoder.encode(DataContainer(data: $0)))
-            }
-        )
-    }
-}
-
-extension Data {
-    func surroundInDataJSONContainer() -> Data {
-        // Surround in a data json container {"data":<existing contents>}
-        var result: [UInt8] = [123, 34, 100, 97, 116, 97, 34, 58, 125]
-        result.insert(contentsOf: self, at: 8)
-        return .init(result)
-//        var bytes = Array(self)
-//        bytes.insert(contentsOf: [123, 34, 100, 97, 116, 97, 34, 58], at: 0)
-//        bytes.append(125)
-//        return .init(bytes)
-    }
-}
